Add empty menu placeholder actions using functional context keys from #6965 (#7038)
# References and relevant issues

Alternative to #6950 that doesn't rely on registering/deregistering
actions but instead uses the mechanism from #6965.

Depends on: #6965

# Description

On macOS, there is no indication by default that a menubar submenu is
empty — one simply has to hover for an unnaturally long time to be sure
that indeed, there is nothing in that submenu. Instead of allowing this
behaviour, apps usually put an "Empty" placeholder in empty submenus
that quickly indicate to the user that there is nothing in this menu.

This is particularly critical for napari users since #7011, because that
PR added a *lot* of menus that are empty if the user has not installed
plugins.

In #6950, I tried to add this issue by:
- registering a do-nothing action
- de-registering it and re-registering it whenever the menus changed,
placing it in all menus that are empty.

The issue with that approach is that de-registering and re-registering
the action was itself changing the menus, which could cause infinite
loops if we weren't super careful — and indeed we did, and worse, we
seemed to do so stochastically.

This PR instead uses just-in-time functional context evaluation to check
whether a menu is empty as it's about to be shown, and show the empty
key if so.

https://github.com/napari/napari/assets/17995243/9b051a68-fe1d-41d9-ad39-b7fcc11d5206

## Trying stuff out

The easiest way to try this out is to just launch `napari` from main and
look for an empty menu e.g. `Layers -> Data`. Then, if you want to see
the placeholder disappear upon action registration, you can run the
following code in the console.

```python
from napari._app_model import get_app
from app_model.types import Action

my_action = Action(
    id='napari.new_action',
    title='On The Fly',
    callback=lambda: napari.utils.notifications.show_info("WOOOOOW"),
    menus=[{'id': 'napari/layers/data'}]
    )

app = get_app()
deregister_action = app.register_action(my_action)
# call deregister_action() to deregister the same action
```

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>